ambulanceserviceprovider/ambulanceserviceprovider/src/main/java/com/example/ambulanceserviceprovider/controller/RequestController.java
package com.example.ambulanceserviceprovider.controller;

import com.example.ambulanceserviceprovider.model.Request;
import com.example.ambulanceserviceprovider.service.RequestService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

        import java.util.List;

@RestController
@RequestMapping("/api/requests")
public class RequestController {
    @Autowired
    private RequestService requestService;

    @PostMapping
    public ResponseEntity<Request> createRequest(@RequestBody Request request) {
        return ResponseEntity.ok(requestService.createRequest(request));
    }

    @GetMapping
    public ResponseEntity<List<Request>> getAllRequests() {
        return ResponseEntity.ok(requestService.getAllRequests());
    }

    // Other endpoints for request management
}